requirements.txt
pynput
pyaudio
wave
playsound==1.2.2
deepgram-sdk==2.12.0
python-dotenv
customtkinter
pillow
pystray
openai>=1.0.0
